Limit the number of characters in a filename
In group chats the numbers are concatenated which could end up in hundreds of characters, sometimes overflowing the allowed number of characters of some systems. This prevent the script from running. That's why I arbitrary added a limit of 100 characters, which is lower that the limit of 255 used by a lot of systems, but that can be discussed.